ssh/tailssh: unify some of the incubator_* GOOS files into incubator.go
In prep for fix for #6888
